Moved utility function from the DNS server package into a common package for sidecar.
