+++ sidecar/dns/server.go
@@ -17,8 +17,6 @@ package dns
 import (
 	"fmt"
 	"net"
-	"net/url"
-	"strconv"
 	"strings"
 
 	"math/rand"
@@ -27,6 +25,7 @@ import (
 
 	"github.com/Sirupsen/logrus"
 	"github.com/amalgam8/amalgam8/pkg/api"
+	"github.com/amalgam8/amalgam8/sidecar/util"
 	"github.com/miekg/dns"
 )
 
@@ -273,7 +272,7 @@ func (s *Server) createRecords(question dns.Question, request, response *dns.Msg
 	extra := make([]dns.RR, 0, 3)
 
 	for _, instance := range instances {
-		ip, port, err := splitHostPort(instance.Endpoint)
+		ip, port, err := util.SplitHostPort(instance.Endpoint)
 		if err != nil {
 			logrus.WithError(err).Warnf("unable to resolve ip address for instance '%s' in DNS query '%s'",
 				instance.ID, question.Name)
@@ -319,70 +318,6 @@ func (s *Server) createRecords(question dns.Question, request, response *dns.Msg
 	response.SetRcode(request, dns.RcodeSuccess)
 	return nil
 
-}
-
-func splitHostPort(endpoint api.ServiceEndpoint) (net.IP, uint16, error) {
-	switch endpoint.Type {
-	case "tcp", "udp":
-		return splitHostPortTCPUDP(endpoint.Value)
-	case "http", "https":
-		return splitHostPortHTTP(endpoint.Value)
-	default:
-		return nil, 0, fmt.Errorf("unsupported endpoint type: %s", endpoint.Type)
-	}
-}
-
-func splitHostPortTCPUDP(value string) (net.IP, uint16, error) {
-	// Assume value is "host:port"
-	host, port, err := net.SplitHostPort(value)
-
-	// Assume value is "host" (no port)
-	if err != nil {
-		host = value
-		port = "0"
-	}
-
-	ip := net.ParseIP(host)
-	if ip == nil {
-		return nil, 0, fmt.Errorf("could not parse '%s' as ip:port", value)
-	}
-
-	portNum, err := strconv.Atoi(port)
-	if err != nil {
-		return nil, 0, err
-	}
-
-	return ip, uint16(portNum), nil
-}
-
-func splitHostPortHTTP(value string) (net.IP, uint16, error) {
-	isHTTP := strings.HasPrefix(value, "http://")
-	isHTTPS := strings.HasPrefix(value, "https://")
-	if !isHTTPS && !isHTTP {
-		value = "http://" + value
-		isHTTP = true
-	}
-
-	parsedURL, err := url.Parse(value)
-	if err != nil {
-		return nil, 0, err
-	}
-
-	ip, port, err := splitHostPortTCPUDP(parsedURL.Host)
-	if err != nil {
-		return nil, 0, err
-	}
-
-	// Use default port, if not specified
-	if port == 0 {
-		if isHTTP {
-			port = 80
-		} else if isHTTPS {
-			port = 443
-		}
-	}
-
-	return ip, port, nil
 }
 
 func createARecord(name string, ip net.IP) *dns.A {
